feat(web): add notification list
