+++ apps/web/package.json
@@ -17,6 +17,7 @@
 		"@tanstack/react-query": "4.26.1",
 		"@tanstack/react-query-devtools": "4.26.1",
 		"common": "workspace:*",
+		"moment": "2.29.4",
 		"next": "13.2.4",
 		"openapi-typescript-fetch": "1.1.3",
 		"react": "18.2.0",
@@ -36,6 +37,7 @@
 		"@storybook/react": "7.0.0-rc.2",
 		"@storybook/testing-library": "0.0.14-next.1",
 		"@svgr/webpack": "6.5.1",
+		"@types/moment": "2.13.0",
 		"@types/node": "18.15.0",
 		"@types/react": "18.0.28",
 		"@types/react-dom": "18.0.11",
