+++ apps/web/src/components/dashboard/DashboardHeader/DashboardHeader.tsx
@@ -1,4 +1,4 @@
-import { NotificationsMenu } from './NotificationsMenu';
+import { NotificationsMenu } from './NotificationMenu/NotificationsMenu';
 import { UserMenu } from './UserMenu/UserMenu';
 
 export const HEADER_MENU_TOP = 55;
